requirements.txt
altair>=5.0
matplotlib>=3.8
numpy>=2.0
pandas>=2.2
pydantic>=2.7
pyyaml>=6.0
regex>=2024.5
rich>=13.7
scipy>=1.11
torch>=2.1
transformers>=4.44
